refactor: 자동차 이름과 시도 횟수 유효성 검사 함수를 validator로 통합
